+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.16.1] - Unreleased
+### Fixed
+- Fix issue [#85](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/85) when checking the validity of a 
+  connection to AstraDB with `CassandraConnection.isValid(int)`.
+
 ## [4.16.0] - 2025-08-09
 ### Added
 - Add new specific protocol `jdbc:cassandra:astra` for support of Datastax AstraDB without specifying a secure connect
@@ -370,6 +375,7 @@ For this version, the changelog lists the main changes comparatively to the late
 - Fix logs in `CassandraConnection` constructor.
 
 [original project]: https://github.com/adejanovski/cassandra-jdbc-wrapper/
+[4.16.1]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.16.0...v4.16.1
 [4.16.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.15.0...v4.16.0
 [4.15.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.14.0...v4.15.0
 [4.14.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.13.1...v4.14.0